pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "website to write blog posts"
authors = [{ name = "bob jiang", email = "bobjiang4321@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "bleach>=6.3.0",
    "django>=6.0.2",
    "django-ckeditor>=6.7.3",
    "django-cors-headers>=4.9.0",
    "django-haystack>=3.3.0",
    "djangorestframework>=3.16.1",
    "jieba>=0.42.1",
    "pillow>=12.1.0",
    "pymysql>=1.1.2",
    "pyyaml>=6.0.3",
    "tinycss2>=1.5.1",
    "uwsgi>=2.0.31",
    "whoosh>=2.7.4",
]
